sdk: refactor position status calculation
